+++ src/main/java/com/dwarfeng/ftp/handler/FtpHandler.java
@@ -3,6 +3,9 @@ package com.dwarfeng.ftp.handler;
 import com.dwarfeng.subgrade.stack.exception.HandlerException;
 import com.dwarfeng.subgrade.stack.handler.Handler;
 
+import java.io.InputStream;
+import java.io.OutputStream;
+
 /**
  * FTP 处理器。
  *
@@ -57,6 +60,35 @@ public interface FtpHandler extends Handler {
      * @throws HandlerException 处理器异常。
      */
     byte[] getFileContent(String[] filePaths, String fileName) throws HandlerException;
+
+    /**
+     * 通过流的形式存储文件。
+     *
+     * <p>
+     * FtpHandler 本身不维护流的生命周期，请在调用该方法前妥善启动流，并在调用该方法结束后妥善关闭流。
+     *
+     * @param filePaths 文件夹路径。<br>
+     *                  路径从根文件出发，一直到达最后一个文件夹，所有文件夹按照顺序组成数组。
+     * @param fileName  文件的名称。
+     * @param in        文件的输入流。
+     * @throws HandlerException 处理器异常。
+     */
+    void storeFileByStream(String[] filePaths, String fileName, InputStream in) throws HandlerException;
+
+    /**
+     * 通过流的形式获取文件。
+     *
+     * <p>
+     * FtpHandler 本身不维护流的生命周期，请在调用该方法前妥善启动流，并在调用该方法结束后妥善关闭流。
+     *
+     * @param filePaths 文件夹路径。<br>
+     *                  路径从根文件出发，一直到达最后一个文件夹，所有文件夹按照顺序组成数组。
+     * @param fileName  文件的名称
+     * @param out       待写入文件内容的输出流。
+     * @throws HandlerException 处理器异常。
+     */
+    void getFileContentByStream(String[] filePaths, String fileName, OutputStream out)
+            throws HandlerException;
 
     /**
      * 删除文件。